% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGCA.decorrelate.R
\name{TGCA.decorrelate}
\alias{TGCA.decorrelate}
\title{Decorrelation Z-scores with estimated phenotypic correlations}
\usage{
TGCA.decorrelate(tstat, N, MAF, maf.cut = 5e-04, eigen.cut = 0.9)
}
\arguments{
\item{tstat}{A data frame of GWAS Z-scores across K phenotypes of M SNPs.}

\item{N}{A vector of the sample size of K phenotypes.}

\item{MAF}{A vector of minor alle frequency of M SNPs.}

\item{maf.cut}{A value defining the low-MAF cutoff, where the SNPs with lower MAF values will be used to estimate the phenotypic correlations. Default 5e-4.}

\item{eigen.cut}{A value defining the proportion of information captured by the decorrelated eigenvectors. Default 0.9.}
}
\value{
A list of two elements, where \code{z.decorrelated} contains a matrix of decorrelated Z-scores corresponding to the M SNPs, and \code{cor.pheno} gives the estimated phenotypic correlation matrix.
}
\description{
The function re-weights GWAS Z-scores acorss multiple phenotypes by sample sizes and adjusts their phenotypic correlation before running TGCA.
}
\note{
You can also use this fuction to re-weight GWAS Z-scores acorss multiple phenotypes by sample size and adjust their phenotypic correlation.
}
\examples{
\dontrun{
data(tgca)
decor <- TGCA.decorrelate(tstat, MAF, N)
image(cor(decor$z.decorrelated))
image(decor$cor.pheno)
}

}
\references{
Shen, X., Li, T., Ning, Z. (2020). Improved estimation of phenotypic correlations using summary association statistics. bioRxiv.
}
\author{
Ting Li, Xia Shen
}
