% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgca.data.R
\name{tgca}
\alias{tgca}
\title{Example data}
\usage{
tgca()
}
\arguments{
\item{MAF}{A vector of minor allele frequencies for M SNPs.}

\item{tstat}{A matrix of Z-scores from GWAS of K traits for the M SNPs.}

\item{N}{A vector of sample sizes of the K phenotypes.}
}
\description{
The data contains the objects required for running the examples
}
\examples{
\dontrun{
data(tgca)
ls()
}

}
\references{
Li, T., ..., Ning, Z. & Shen, X. Total genetic contribution assessment of the human genome. Submitted (2020).
}
\author{
Ting Li, Xia Shen
}
