% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGCA.scan.R
\name{TGCA.scan}
\alias{TGCA.scan}
\title{Total genetic contribution assessment}
\usage{
TGCA.scan(zmat)
}
\arguments{
\item{zmat}{A matrix of decorrelated Z-scores corresponding to M SNPs.}
}
\value{
A matrix of TGCA results for M SNPs with estimates, standard errors, and p-values of TGCA theta, pi0, pi-, pi+, mu-, mu+, sigma- ,and sigma+.
}
\description{
The function estimates the total genetic contribution of a series of SNPs.
}
\note{
you can also use this fuction to re-weight GWAS z-scores acorss multiple phenotypes by sample size and adjust their phenotypic correlation.
}
\examples{
\dontrun{
data(tgca)
decor <- TGCA.decorrelate(tstat, MAF, N)
idx <- which(MAF < 5e-4)
zmat <- decor$z.decorrelated[-idx,]
res <- TGCA.scan(zmat)
}

}
\references{
Li, T., ..., Ning, Z. & Shen, X. Total genetic contribution assessment of the human genome. Submitted (2020).
}
\author{
Ting Li, Xia Shen
}
